(chore) reorganized NST feature for better refactoring later on
