+++ csdt/snap.html
@@ -49,9 +49,9 @@
 		<script src="../src/FileSaver.min.js?version=2019-06-27"></script>
 
 		<script src="./core/jquery.min.js"></script>
-		<script src="./ast/bundle.js"></script>
+		<script src="./features/nst/lib/bundle.js"></script>
 		<script src="./core/tutorials.js"></script>
-		<script src="./core/visualizer.js"></script>
+		<!-- <script src="./core/visualizer.js"></script> -->
 		<script src="./core/main.js?newversion" type="module"></script>
 		<script>
 			var world;
